refactor: repair dependency
