+++ app/src/main/java/kosenda/makecolor/navigation/Navigation.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor
+package kosenda.makecolor.navigation
 
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalAnimationApi
@@ -26,6 +26,7 @@ import kosenda.makecolor.core.model.data.ColorItem
 import kosenda.makecolor.core.ui.data.ColorIndex
 import kosenda.makecolor.core.ui.data.SplitColorParam
 import kosenda.makecolor.core.ui.feature.common.SelectColorParam
+import kosenda.makecolor.feature.display.screen.CategoryDetailScreen
 import kosenda.makecolor.feature.display.screen.ColorDetailScreen
 import kosenda.makecolor.feature.display.screen.DataScreen
 import kosenda.makecolor.feature.display.screen.FullColorScreen
@@ -34,7 +35,6 @@ import kosenda.makecolor.feature.display.screen.GradationScreen
 import kosenda.makecolor.feature.display.screen.SelectColorScreen
 import kosenda.makecolor.feature.display.screen.SplitColorScreen
 import kosenda.makecolor.feature.display.screen.SplitScreen
-import kosenda.makecolor.feature.edit.CategoryDetailScreen
 import kosenda.makecolor.feature.edit.RegisterScreen
 import kosenda.makecolor.feature.info.InfoDialog
 import kosenda.makecolor.feature.makecolor.screen.InputTextScreen
@@ -44,6 +44,7 @@ import kosenda.makecolor.feature.makecolor.screen.PictureScreen
 import kosenda.makecolor.feature.makecolor.screen.RandomScreen
 import kosenda.makecolor.feature.makecolor.screen.SeekbarScreen
 import kosenda.makecolor.feature.settings.SettingScreen
+import kosenda.makecolor.ui.GoogleAd
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
