+++ app/src/main/java/kosenda/makecolor/ui/DrawerItem.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor
+package kosenda.makecolor.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -21,10 +21,12 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import kosenda.makecolor.R
 import kosenda.makecolor.core.ui.feature.common.LocalIsDark
 import kosenda.makecolor.core.ui.feature.theme.changeBrightness
 import kosenda.makecolor.core.ui.feature.theme.contentBrush
 import kosenda.makecolor.core.ui.feature.theme.primaryBrush
+import kosenda.makecolor.navigation.NavigationItems
 
 @Composable
 fun DrawerItem(
