+++ core/model/src/main/java/kosenda/makecolor/core/model/data/RandomType.kt
@@ -3,6 +3,6 @@ package kosenda.makecolor.core.model.data
 enum class RandomType(val index: Int) {
     NOT_SPECIFIED(index = 0),
     PASTEL(index = 1),
-    Vivid(index = 2),
+    VIVID(index = 2),
     BLACK_AND_WHITE(index = 3),
 }
