+++ core/ui/src/main/java/kosenda/makecolor/core/ui/feature/common/dialog/NewCategoryDialogContent.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.feature.edit
+package kosenda.makecolor.core.ui.feature.common.dialog
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -31,6 +31,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kosenda.makecolor.core.model.data.Category
+import kosenda.makecolor.core.ui.R
 import kosenda.makecolor.core.ui.feature.common.button.CustomTextButton
 import kosenda.makecolor.core.ui.feature.common.button.TonalButton
 import kosenda.makecolor.core.ui.feature.common.card.TitleCard
