+++ feature/display/src/main/java/kosenda/makecolor/feature/display/screen/CategoryDetailScreen.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.feature.edit
+package kosenda.makecolor.feature.display.screen
 
 import android.widget.Toast
 import androidx.compose.foundation.background
@@ -30,10 +30,15 @@ import androidx.compose.ui.unit.dp
 import kosenda.makecolor.core.model.data.Category
 import kosenda.makecolor.core.ui.feature.common.button.DeleteButton
 import kosenda.makecolor.core.ui.feature.common.button.TonalButton
+import kosenda.makecolor.core.ui.feature.common.dialog.NewCategoryDialog
 import kosenda.makecolor.core.ui.feature.common.topbar.BackTopBar
 import kosenda.makecolor.core.ui.feature.theme.MakeColorTheme
 import kosenda.makecolor.core.util.getNameIfNoAlias
 import kosenda.makecolor.core.util.hexToColor
+import kosenda.makecolor.feature.display.R
+import kosenda.makecolor.feature.display.viewmodel.CategoryDetailViewModel
+import kosenda.makecolor.feature.display.viewmodel.CategoryDetailViewModelImpl
+import kosenda.makecolor.feature.display.viewmodel.PreviewCategoryDetailViewModel
 import kosenda.makecolor.feature.preview.PreviewSurface
 
 @Composable
